refactor(front): migrate Icon to ui/Icon
Drop `Icon` from _dsfr in 5 consumers; route through the project's
~/components/ui/Icon wrapper. Map size="1x" -> "md", "xl" -> "lg";
drop iconPosition/verticalAlign - handled by parent layout or by
ui/Icon's centered span.

Also replace the ExtendedToggle default `fr-icon-check-fill` (not a
real DSFR icon class) with `fr-icon-check-line` so the new typed
Icon prop accepts it.
