+++ frontend/src/components/ExtendedToggle/ExtendedToggle.tsx
@@ -1,13 +1,14 @@
+import classNames from 'classnames';
+import { useEffect, useMemo } from 'react';
+import { v4 as uuidv4 } from 'uuid';
+import Icon from '~/components/ui/Icon';
+
 import type {
   FrIconClassName,
   RiIconClassName
 } from '@codegouvfr/react-dsfr/src/fr/generatedFromCss/classNames';
 import type { ToggleSwitchProps } from '@codegouvfr/react-dsfr/ToggleSwitch';
-import classNames from 'classnames';
-import { useEffect, useMemo } from 'react';
-import { v4 as uuidv4 } from 'uuid';
 
-import { Icon } from '../_dsfr';
 
 import styles from './extended-toggle.module.scss';
 
@@ -49,7 +50,7 @@ function ExtendedToggle(props: Props) {
         'data-fr-unchecked-label': 'Désactivé'
       };
 
-  const icon = props.iconId ?? 'fr-icon-check-fill';
+  const icon = props.iconId ?? 'fr-icon-check-line';
 
   return (
     <div className={classes}>
@@ -61,12 +62,7 @@ function ExtendedToggle(props: Props) {
           className={classNames(styles.input)}
           id={id}
         />
-        <Icon
-          className={styles.icon}
-          iconPosition="center"
-          name={icon}
-          size="xs"
-        />
+        <Icon className={styles.icon} name={icon} size="xs" />
       </div>
       <label className={styles.label} htmlFor={id} {...checkboxLabels}>
         {props.label}
